ic_agent
Module agent

The main Agent module. Contains the Agent type and all associated structures.

Re-exports

pub use agent_error::AgentError;

Modules

agent_error

Errors that can occur when using the replica agent.

http_transport

Transport implementations.

signed

Types representing signed messages.

status

Types for interacting with the status endpoint of a replica. See Status for details.

Structs

Agent

A low level Agent to make calls to a Replica endpoint.

AgentBuilder

A builder for an Agent.

AgentConfig

A configuration for an agent.

ApiBoundaryNode

API boundary node, which routes /api calls to IC replica nodes.

Envelope

The authentication envelope, containing the contents and their signature.

NonceFactory

A Factory for nonce blobs.

QueryBuilder

A Query Request Builder.

RejectResponse

An IC execution error received from the replica.

ReplyResponse

A successful reply to a canister call.

UpdateBuilder

An Update Request Builder.

UpdateCall

An in-flight canister update call. Useful primarily as a Future.

Enums

CallResponse

The response from a request to the call endpoint.

EnvelopeContent

The content of an IC ingress message, not including any signature information.

RejectCode

See the interface spec.

RequestStatusResponse

The response of /api/v2/canister/<effective_canister_id>/read_state with request_status request type.

Traits

NonceGenerator

An interface for generating nonces.

Transport

A facade that connects to a Replica and does requests. These requests can be of any type (does not have to be HTTP). This trait is to inverse the control from the Agent over its connection code, and to resolve any direct dependencies to tokio or HTTP code from this crate.

Functions

signed_query_inspect

Inspect the bytes to be sent as a query Return Ok only when the bytes can be deserialized as a query and all fields match with the arguments

signed_request_status_inspect

Inspect the bytes to be sent as a request_status Return Ok only when the bytes can be deserialized as a request_status and all fields match with the arguments

signed_update_inspect

Inspect the bytes to be sent as an update Return Ok only when the bytes can be deserialized as an update and all fields match with the arguments