ic_agent::agent_error
Enum AgentError

pub enum AgentError {

    InvalidReplicaUrl(String),
    TimeoutWaitingForResponse(),
    SigningError(String),
    InvalidCborData(Error),
    CannotCalculateRequestId(RequestIdError),
    CandidError(Box<dyn Send + Sync + Error>),
    UrlParseError(ParseError),
    InvalidMethodError(InvalidMethod),
    PrincipalError(PrincipalError),
    CertifiedReject(RejectResponse),
    UncertifiedReject(RejectResponse),
    HttpError(HttpErrorPayload),
    InvalidReplicaStatus,
    RequestStatusDoneNoReply(String),
    MessageError(String),
    Leb128ReadError(Error),
    Utf8ReadError(Utf8Error),
    LookupPathAbsent(Vec<Label>),
    LookupPathUnknown(Vec<Label>),
    LookupPathError(Vec<Label>),
    InvalidRequestStatus(Vec<Label>, String),
    CertificateVerificationFailed(),
    QuerySignatureVerificationFailed,
    CertificateNotAuthorized(),
    CertificateOutdated(Duration),
    CertificateHasTooManyDelegations,
    MissingSignature,
    MalformedSignature,
    MalformedPublicKey,
    TooManySignatures {
        had: usize,
        needed: usize,
    },
    DerKeyLengthMismatch {
        expected: usize,
        actual: usize,
    },
    DerPrefixMismatch {
        expected: Vec<u8>,
        actual: Vec<u8>,
    },
    NoRootKeyInStatus(Status),
    WalletCallFailed(String),
    WalletError(String),
    WalletUpgradeRequired(String),
    MissingReplicaTransport(),
    ResponseSizeExceededLimit(),
    TransportError(Box<dyn Error + Send + Sync>),
    CallDataMismatch {
        field: String,
        value_arg: String,
        value_cbor: String,
    },
    InvalidRejectCode(InvalidRejectCodeError),
    RouteProviderError(String),
    InvalidHttpResponse(String),

}

An error that occurred when using the agent.

Variants

InvalidReplicaUrl(String)

The replica URL was invalid.

TimeoutWaitingForResponse()

The request timed out.

SigningError(String)

An error occurred when signing with the identity.

InvalidCborData(Error)

The data fetched was invalid CBOR.

CannotCalculateRequestId(RequestIdError)

There was an error calculating a request ID.

CandidError(Box<dyn Send + Sync + Error>)

There was an error when de/serializing with Candid.

UrlParseError(ParseError)

There was an error parsing a URL.

InvalidMethodError(InvalidMethod)

The HTTP method was invalid.

PrincipalError(PrincipalError)

The principal string was not a valid principal.

CertifiedReject(RejectResponse)

The subnet rejected the message.

UncertifiedReject(RejectResponse)

The replica rejected the message. This rejection cannot be verified as authentic.

HttpError(HttpErrorPayload)

The replica returned an HTTP error.

InvalidReplicaStatus

The status endpoint returned an invalid status.

RequestStatusDoneNoReply(String)

The call was marked done, but no reply was provided.

MessageError(String)

A string error occurred in an external tool.

Leb128ReadError(Error)

There was an error reading a LEB128 value.

Utf8ReadError(Utf8Error)

A string was invalid UTF-8.

LookupPathAbsent(Vec<Label>)

The lookup path was absent in the certificate.

LookupPathUnknown(Vec<Label>)

The lookup path was unknown in the certificate.

LookupPathError(Vec<Label>)

The lookup path did not make sense for the certificate.

InvalidRequestStatus(Vec<Label>, String)

The request status at the requested path was invalid.

CertificateVerificationFailed()

The certificate verification for a read_state call failed.

QuerySignatureVerificationFailed

The signature verification for a query call failed.

CertificateNotAuthorized()

The certificate contained a delegation that does not include the effective_canister_id in the canister_ranges field.

CertificateOutdated(Duration)

The certificate was older than allowed by the ingress_expiry.

CertificateHasTooManyDelegations

The certificate contained more than one delegation.

MissingSignature

The query response did not contain any node signatures.

MalformedSignature

The query response contained a malformed signature.

MalformedPublicKey

The read-state response contained a malformed public key.

TooManySignatures

The query response contained more node signatures than the subnet has nodes.

Fields

had: usize

The number of provided signatures.

needed: usize

The number of nodes on the subnet.

DerKeyLengthMismatch

There was a length mismatch between the expected and actual length of the BLS DER-encoded public key.

Fields

expected: usize

The expected length of the key.

actual: usize

The actual length of the key.

DerPrefixMismatch

There was a mismatch between the expected and actual prefix of the BLS DER-encoded public key.

Fields

expected: Vec<u8>

The expected key prefix.

actual: Vec<u8>

The actual key prefix.

NoRootKeyInStatus(Status)

The status response did not contain a root key.

WalletCallFailed(String)

The invocation to the wallet call forward method failed with an error.

WalletError(String)

The wallet operation failed.

WalletUpgradeRequired(String)

The wallet canister must be upgraded. See dfx wallet upgrade

MissingReplicaTransport()

The transport was not specified in the AgentBuilder.

ResponseSizeExceededLimit()

The response size exceeded the provided limit.

TransportError(Box<dyn Error + Send + Sync>)

An unknown error occurred during communication with the replica.

CallDataMismatch

There was a mismatch between the expected and actual CBOR data during inspection.

Fields

field: String

The field that was mismatched.

value_arg: String

The value that was expected to be in the CBOR.

value_cbor: String

The value that was actually in the CBOR.

InvalidRejectCode(InvalidRejectCodeError)

The rejected call had an invalid reject code (valid range 1..5).

RouteProviderError(String)

Route provider failed to generate a url for some reason.

InvalidHttpResponse(String)

Invalid HTTP response.