ic_agent
Module identity

Types and traits dealing with identity across the Internet Computer.

Structs

AnonymousIdentity

The anonymous identity.

BasicIdentity

A Basic Identity which sign using an ED25519 key pair.

DelegatedIdentity

An identity that has been delegated the authority to authenticate as a different principal.

Delegation

A delegation from one key to another.

Prime256v1Identity

A cryptographic identity based on the Prime256v1 elliptic curve.

Secp256k1Identity

A cryptographic identity based on the Secp256k1 elliptic curve.

Signature

A cryptographic signature, signed by an Identity.

SignedDelegation

A Delegation that has been signed by an Identity.

Enums

PemError

An error happened while reading a PEM file.

Traits

Identity

An Identity produces Signatures for requests or delegations. It knows or represents the Principal of the sender.